requirements.txt
duckdb
pandas
pytest
pytest-mock
pyyaml
